requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-bootstrap-icons==0.8.1
django-bootstrap-v5==1.0.11
django-cleanup==6.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
gunicorn==20.1.0
idna==3.3
Pillow==9.2.0
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.11
